requirements.txt
openai
python-dotenv
pydantic
fastapi
uvicorn
websockets
